refactor: some improvements

+++ algorithms/introsort.c
@@ -29,6 +29,6 @@ static void introsort_rec(int a[], size_t lo, size_t hi, size_t depth_limit)
 
 void introsort(int a[], size_t length)
 {
-    size_t depth_limit = 2 * floor(log2(length));
+    size_t depth_limit = (size_t)(2 * floor(log2((double)length)));
     introsort_rec(a, 0, !length ? 0 : length - 1, depth_limit);
 }
